requirements.txt
alembic==1.18.4
aniso8601==10.0.1
blinker==1.9.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.2.2
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
passlib==1.7.4
pillow==12.1.1
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.12.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
reportlab==4.4.10
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.1.6
